requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
trafilatura==1.12.0
unstructured==0.15.7
langchain==0.3.2
langchain-community==0.3.1
faiss-cpu==1.8.0
sentence-transformers==3.0.1
rank-bm25==0.2.2
python-dotenv==1.0.1
PyYAML==6.0.2
openai==1.43.0
